clinical_rag/requirements.txt
faiss-cpu==1.7.4
sentence-transformers==2.7.0
torch==2.2.2
transformers==4.41.2

fastapi==0.110.0
uvicorn==0.29.0

pypdf==4.2.0
numpy==1.26.4
scikit-learn==1.4.2
tqdm==4.66.4

openai==1.30.1

gTTS==2.5.1

SpeechRecognition==3.10.0
pyaudio==0.2.14
